// src/module/telegram/types/callbackActions.enum.ts
export enum ActionTitles {
	FIND_ROOMMATE = '🔍 Найти соседа',
	CREATE_PROFILE = '📝 Создать анкету',
	EDIT_PROFILE = '✏️ Редактировать анкету',
	SHOW_PROFILE = '👁️ Просмотреть анкету',
	BACK_TO_MENU = '🔙 Главное меню',
	LIKE_PROFILE = '❤️ Поставить лайк',
	DEACTIVATE_PROFILE = '🚫 Деактивировать анкету',
	REACREATE_PROFILE = '♻️ Пересоздать анкету',
	SKIP_PROFILE = '⏭️ Пропустить анкету'
}

export enum CallbackActions {
	FIND_ROOMMATE = 'ACTION_FIND_ROOMMATE',
	CREATE_PROFILE = 'ACTION_CREATE_PROFILE',
	EDIT_PROFILE = 'ACTION_EDIT_PROFILE',
	SHOW_PROFILE = 'ACTION_SHOW_PROFILE',
	BACK_TO_MENU = 'ACTION_BACK_TO_MENU',
	LIKE_PROFILE = 'ACTION_LIKE_PROFILE',
	DEACTIVATE_PROFILE = 'ACTION_DEACTIATE_PROFILE',
	REACREATE_PROFILE = 'ACTION_REACREATE_PROFILE',
	SKIP_PROFILE = 'ACTION_SKIP_PROFILE'
}